front: move "show details" feature from TimetableToolbar to TimetableBoardWrapper menu
