+++ front/public/locales/en/operational-studies.json
@@ -60,7 +60,6 @@
       "electricalProfileWithId": "Electrical profiles : ID{{id}}",
       "electricalProfileWithName": "Electrical profiles : {{name}} | ID{{id}}"
     },
-    "displayTrainsWithDetails": "Mask/unmask trains detail",
     "editScenario": "Edit scenario",
     "electricalProfileSet": "Electrical profiles set",
     "errorMessages": {
@@ -70,7 +69,7 @@
     "filterPlaceholder": "Filter, tags",
     "infraLoadingState": "{{step}}/3",
     "infrastructure": "Infrastructure",
-    "lessDetails": "less details",
+    "lessDetails": "Display less details",
     "loadingMacroEditor": "Loading macroscopic editor…",
     "macroEditor": {
       "intervalXmin": "Interval {{minutes}}min",
@@ -124,7 +123,7 @@
     },
     "macroscopic": "MACRO",
     "microscopic": "MICRO",
-    "moreDetails": "more details",
+    "moreDetails": "Display more details",
     "noElectricalProfileSet": "No electrical profile",
     "occurrenceMenu": {
       "delete": "Delete",
